delete obsolete calc_damage endpoint and also pull out routing logic
